// frontend/src/api/generated/models/Tag.ts
export type Tag = { type: 'integer' } & IntegerTag | { type: 'integerSet' } & IntegerSetTag | { type: 'keyOnly' } & KeyOnlyTag | { type: 'real' } & RealTag | { type: 'realSet' } & RealSetTag | { type: 'text' } & TextTag | { type: 'textSet' } & TextSetTag;